requirements.txt
Django==6.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-redis==6.0.0
python-dotenv==1.2.1
psycopg2-binary==2.9.11
bcrypt==5.0.0
Pillow==12.1.1
requests==2.32.5
redis==7.2.1
gunicorn==21.2.0
flake8==6.1.0
